refactor: move Supabase client to lib/supabase/client.ts
